write_active_jobs, configure print scaling, and remove pragmas
