+++ tests/test_get_active_jobs.py
@@ -5,9 +5,11 @@ import unittest
 
 import pandas as pd
 
+
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 
-from get_active_jobs import export_active_jobs
+from get_active_jobs import _parse_datetime  # type: ignore[reportPrivateUsage]
+from write_active_jobs import export_active_jobs
 
 FIXTURE = "tests/fixtures/test_operations.xlsx"
 
@@ -45,22 +47,24 @@ class TestGetActiveJobs(unittest.TestCase):
         with tempfile.NamedTemporaryFile(suffix=".xlsx", delete=False) as tmp:
             tmp_path: str = tmp.name
 
+        out_path: str = ""
         try:
             with pd.ExcelWriter(tmp_path) as writer:  # type: ignore[reportUnknownVariableType]
                 df1.to_excel(writer, sheet_name="Sheet1", index=False, header=False)  # type: ignore[reportUnknownMemberType, reportUnknownArgumentType]
                 df2.to_excel(writer, sheet_name="Sheet2", index=False, header=False)  # type: ignore[reportUnknownMemberType, reportUnknownArgumentType]
 
             with tempfile.NamedTemporaryFile(suffix=".xlsx", delete=False) as out:
-                out_path: str = out.name
+                out_path = out.name
             export_active_jobs(tmp_path, "2026-03-12", out_path)
             result: pd.DataFrame = pd.read_excel(out_path, header=5)  # type: ignore[reportUnknownMemberType]
             # No active jobs: 5 rows (sub-header, fallback, blanks, sign)
             self.assertEqual(len(result), 5)
             fallback = str(result.iloc[1]["MACHINE"]).strip()
             self.assertEqual(fallback, "No active jobs scheduled for this shift")
-            os.remove(out_path)
         finally:
             os.remove(tmp_path)
+            if out_path and os.path.exists(out_path):
+                os.remove(out_path)
 
     # --- export_active_jobs ---
 
@@ -90,6 +94,16 @@ class TestGetActiveJobs(unittest.TestCase):
         finally:
             os.remove(tmp_path)
 
+    # --- _parse_datetime branch coverage (no pragma: no cover) ---
 
-if __name__ == "__main__":
-    unittest.main()  # pragma: no cover
+    def test_parse_datetime_returns_none_for_nan(self) -> None:
+        """Cover the pd.isna branch in _parse_datetime."""
+        row = pd.Series([None] * 12)
+        result = _parse_datetime(row, 0)
+        self.assertIsNone(result)
+
+    def test_parse_datetime_returns_none_for_bad_value(self) -> None:
+        """Cover the Exception branch in _parse_datetime."""
+        row = pd.Series(["not-a-date"] * 12)
+        result = _parse_datetime(row, 0)
+        self.assertIsNone(result)
